fixed os specific aliases

+++ aliases.zsh
@@ -4,21 +4,32 @@
 alias system-show-space="df -h"
 alias system-show-space-ncdu="sudo ncdu /"
 alias system-show-my-ip="ip add | grep -o \"inet 192.168.[0-9][0-9]\?[0-9]\?.[0-9][0-9]\?[0-9]\?\" | grep -o \"192.168.[0-9][0-9]\?[0-9]\?.[0-9][0-9]\?[0-9]\?\""
-alias system-freeup-space="sudo pacman -Sc && sudo pacman -Qdt && sudo pacman -Rns $(pacman -Qtdq) && sudo journalctl --vacuum-size=50M && yay -Scc && sudo paccache -rvk0"
-alias system-update="sudo pacman -Syu && yay && flatpak upgrade"
+alias system-show-my-ip-ifconfig="ifconfig | grep -o \"inet 192.168.[0-9][0-9]\?[0-9]\?.[0-9][0-9]\?[0-9]\?\" | grep -o \"192.168.[0-9][0-9]\?[0-9]\?.[0-9][0-9]\?[0-9]\?\""
 
+
+if [[ $OS == "Linux" ]]
+then
+    # Manjaro
+    alias system-freeup-space="sudo pacman -Sc && sudo pacman -Qdt && sudo pacman -Rns $(pacman -Qtdq) && sudo journalctl --vacuum-size=50M && yay -Scc && sudo paccache -rvk0"
+    alias system-update="sudo pacman -Syu && yay && flatpak upgrade"
+    alias pacman-search="pacman -Ss"
+    alias pacman-search-installed="pacman -Qs"
+    alias pacman-ls-installed="pacman -Qi | egrep '^(Name|Installed)' | cut -f2 -d':' | paste - - | column -t | sort -nrk 2 | grep MiB | less"
+    alias pacman-install="sudo pacman -Syu"
+    alias pacman-uninstall="sudo pacman -Rns"
+
+    echo $OS "aliases loaded"
+elif [[ $OS == "OSX" ]]
+
+    echo $OS "aliases loaded"
+then
+
+fi
 
 # Configs
-alias zsh-reload="exec zsh && source ~/.zshrc"
+alias zsh-reload="source ~/.zshrc"
 alias tmux-reload="tmux source ~/.tmux.conf"
 alias zsh-config="code ~/configs/."
-
-# Manjaro
-alias pacman-search="pacman -Ss"
-alias pacman-search-installed="pacman -Qs"
-alias pacman-ls-installed="pacman -Qi | egrep '^(Name|Installed)' | cut -f2 -d':' | paste - - | column -t | sort -nrk 2 | grep MiB | less"
-alias pacman-install="sudo pacman -Syu"
-alias pacman-uninstall="sudo pacman -Rns"
 
 # VPN
 ## Wireguard
@@ -28,7 +39,8 @@ alias vpn-disconnect-pi="wg-quick down wg0"
 # Docker
 alias docker-show-used-spase="docker system df"
 alias docker-remove-all-stopped-containers="docker container prune"
-alias docker-remove-all-images="docker image rm $(docker image ls -q)"
+alias docker-remove-all-images="docker image prune -a"
+alias docker-remove-all-cache="docker builder prune -a"
 
 # Python
 alias python=/usr/local/bin/python3
